third_party/blink/web_tests/virtual/scoped-custom-element-registry/external/wpt/custom-elements/registries/scoped-registry-append-does-not-upgrade-expected.txt

This is a testharness.js-based test.
[FAIL] Connecting a custom element candiate with a scoped custom element registry does not change the registry
  assert_equals: expected object "[object CustomElementRegistry]" but got object "[object CustomElementRegistry]"
Harness: the test ran to completion.
